requirements.txt
blas=1.0=mkl
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=35.0.0=py38ha5dfef3_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.11.0=pypi_0
decorator=4.4.2=py_0
ffmpeg=4.2.2=h20bf706_0
fonttools=4.29.1=pypi_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
googledrivedownloader=0.4=pyhd3deb0d_1
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=h14aa051_20
libgfortran4=7.5.0=h14aa051_20
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.5.1=pyhd8ed1ab_0
ninja=1.10.2=py38hd09550d_3
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.2.3=py38h51da96c_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
pycparser=2.21=pyhd8ed1ab_0
pyg=2.0.1=py38_torch_1.8.0_cu111
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pyhd8ed1ab_1
python_abi=3.8=2_cp38
pytorch=1.8.2=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-cluster=1.5.9=py38_torch_1.8.0_cu111
pytorch-scatter=2.0.8=py38_torch_1.8.0_cu111
pytorch-sparse=0.6.12=py38_torch_1.8.0_cu111
pytorch-spline-conv=1.2.1=py38_torch_1.8.0_cu111
pytz=2021.3=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
scikit-learn=1.0.2=py38h51133e4_1
scipy=1.7.3=py38hc147768_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.8.2=py38
torchvision=0.9.2=py38_cu111
tqdm=4.63.0=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h516909a_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
